Reorganize routes
Add routes public.py ,user_functions.py ,users.py
