+++ app/routes/user_functions.py
@@ -1,0 +1,9 @@
+from flask import Blueprint,render_template,redirect,request,flash
+
+user_functions = Blueprint('user_functions',__name__)
+
+"""
+Aqui van a estar las rutas relacionadas con las funciones que 
+puede realizar el usuario como crear prestamos,servicios,etc.
+
+"""
